All events now displayed in each tab
