+++ app/src/main/java/org/hackillinois/android/view/MainActivity.kt
@@ -12,6 +12,7 @@ import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.layout_nav_menu.*
 import org.hackillinois.android.R
 import org.hackillinois.android.view.home.HomeFragment
+import org.hackillinois.android.view.schedule.ScheduleFragment
 
 class MainActivity : AppCompatActivity(), View.OnClickListener {
